apng-processor/requirements.txt
Pillow>=10.0.0
apng>=0.3.4
imageio>=2.31.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
